requirements.txt
Flask
pdfplumber
python-docx
pandas
pytesseract
Pillow
tqdm
gunicorn
